Move codes to test_util.rs, add comments and test
