+++ rust/arrow/src/util/mod.rs
@@ -22,5 +22,4 @@ pub mod integration_util;
 #[cfg(feature = "prettyprint")]
 pub mod pretty;
 pub mod string_writer;
-pub mod test_data_dir;
 pub mod test_util;
